Converted size calculation to use a long
A long was being returned, but the counter variable was an int, which
could lead to overflow.

+++ scifio/src/main/java/io/scif/AbstractImageMetadata.java
@@ -341,7 +341,7 @@ public abstract class AbstractImageMetadata implements ImageMetadata {
    * @see io.scif.ImageMetadata#getSize()
    */
   public long getSize() {
-    int size = 1;
+    long size = 1;
     
     for (AxisType a : axisTypes) {
       size *= getAxisLength(a);
